% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{curlr}
\alias{curlr}
\title{curlr}
\usage{
curlr(
  curl_statement = "",
  url = "",
  endpoint = "http://localhost:5820",
  db = "",
  Username = "",
  myAuth = "",
  VERB = "POST",
  query = "",
  payload = "",
  error_log = file.path("logs", "curlr_last_error.txt"),
  return_json = FALSE,
  show_progress = FALSE,
  show_response = FALSE,
  assign_outcome = FALSE,
  assign_to_env = caller_env()
)
}
\description{
curlr
}
