% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{source_packages}
\alias{source_packages}
\title{Install and load multiple packages at once}
\usage{
source_packages(
  pkgs,
  show_loaded_packages = F,
  show_package_processes = F,
  lib = .libPaths()[[1]],
  repos = "http://cran.us.r-project.org"
)
}
\arguments{
\item{pkgs}{A character vector of package names to install as needed and load to the environment.}

\item{show_loaded_packages}{If TRUE, echo the list of packages loaded.}

\item{show_package_processes}{If TRUE, echo messages about the installation processes.}

\item{repos}{defaults to 'http://cran.us.r-project.org'}
}
\value{
Installs and loads the packages.
}
\description{
Install and load multiple packages at once
}
