% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{sourceR}
\alias{sourceR}
\title{sourceR}
\usage{
sourceR(
  source_dir = "R",
  include_rmd = F,
  exclude_dirs = "",
  exclude_files = "",
  recursive = T,
  exclude_marked_exclude = T,
  exclude_sourceR = T,
  echo_files_sourced = T
)
}
\arguments{
\item{exclude_sourceR}{}
}
\value{

}
\description{
Source scripts selectively, recursively
}
