% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_}
\alias{stardog_}
\title{stardog_}
\usage{
stardog_(q = query, d = db, g = graph, U = Username, e = endpoint)
}
\description{

}
\examples{
stardog_(q = query, d = db, g = graph, U = Username, e = endpoint)
}
