% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_add_namespaces}
\alias{stardog_add_namespaces}
\title{stardog_add_namespaces}
\usage{
stardog_add_namespaces(
  endpoint = "http://localhost:5820",
  db = "",
  Username = "admin",
  input_file,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
stardog_add_namespaces
}
