% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_create_db}
\alias{stardog_create_db}
\title{stardog_create_db}
\usage{
stardog_create_db(db_name, endpoint = "http://localhost:5820")
}
\description{
stardog_create_db
}
\examples{
stardog.create_db("cars")
}
