% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_data_add}
\alias{stardog_data_add}
\title{stardog_data_add}
\usage{
stardog_data_add(
  directory,
  data_file,
  endpoint,
  db,
  Username,
  file_format = "TURTLE",
  remove_all = F
)
}
\description{
stardog_data_add
}
