% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_derive_classes}
\alias{stardog_derive_classes}
\title{stardog_derive_classes}
\usage{
stardog_derive_classes(db, no_instanceless = F)
}
\description{

}
\examples{
stardog_derive_classes(db =  , no_instanceless = F)
}
