% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_http}
\alias{stardog_http}
\title{stardog_http}
\usage{
stardog_http(
  query = "namespaces",
  endpoint = "http://localhost:5820",
  db = "",
  graph = "",
  Username = "admin",
  clear_keys = F,
  httr_method = "GET",
  body = FALSE,
  assign_temp_results = F
)
}
\description{
stardog_http
}
