% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_http_}
\alias{stardog_http_}
\title{stardog_http_}
\usage{
stardog_http_(
  q = query,
  d = db,
  g = graph,
  U = Username,
  e = endpoint,
  m = httr_method,
  b = body
)
}
\description{

}
\examples{
stardog_http_(q = query, d = db, g = graph, U = Username, e = endpoint, m = httr_method, b = body)
}
