% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_list_namespaces}
\alias{stardog_list_namespaces}
\title{stardog_list_namespaces}
\usage{
stardog_list_namespaces(db, assign = FALSE)
}
\description{

}
\examples{
stardog_list_namespaces(db =  , assign = FALSE)
}
