% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_list_user_permissions}
\alias{stardog_list_user_permissions}
\title{stardog_list_user_permissions}
\usage{
stardog_list_user_permissions(user)
}
\description{

}
\examples{
stardog_list_user_permissions(user =  )
}
