% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_virtual_import}
\alias{stardog_virtual_import}
\title{stardog_virtual_import}
\usage{
stardog_virtual_import(
  endpoint = "http://localhost:5820",
  db,
  myAuth = "admin:admin",
  input_file_type = "DELIMITED",
  input_file,
  mapping_file
)
}
\description{
stardog_virtual_import
}
