% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{summon}
\alias{summon}
\title{Summon together the text of many files.}
\usage{
summon(
  files = "",
  file_pattern = ".R$",
  include_rmd = T,
  exclude_files = "",
  exclude_marked_exclude = T,
  exclude_summon = T,
  directory = "R",
  recursive = T,
  summon_as_filetype = "Rmd",
  open_summoned = T
)
}
\arguments{
\item{files}{Optional - A vector of files to summon}

\item{file_pattern}{The file pattern to look for if summoning from a directory, eg the file extension ".R$"}

\item{include_rmd}{Adds "|.Rmd$" to the file_pattern}

\item{exclude_files}{A vector of files to exclude from being summoned}

\item{exclude_marked_exclude}{TRUE to exclude any files in the directory with names ending in _exclude before the file extension}

\item{exclude_summon}{TRUE to exclude the summon.R script file from being summoned}

\item{directory}{The directory to search for matching files}

\item{recursive}{Whether to search recursively through subdirectories}

\item{summon_as_filetype}{What filetype to summon the files into, defaults to .Rmd with a block for each file.}

\item{open_summoned}{TRUE to immediately open the resulting summoned file.}
}
\value{
Creates a file that pastes together the text content of all of the summoned files. If summon_as_filetype = "Rmd" this will be an Rmd file with one block for each file summoned.
}
\description{
Summon together the text of many files.
}
