% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStardog_Functions.R
\name{stardog_virtual_import}
\alias{stardog_virtual_import}
\title{stardog_virtual_import}
\usage{
stardog_virtual_import(directory, csv, sms, endpoint, db, Username)
}
\description{
stardog_virtual_import
}
